Add files via upload
The bookstore manager typically involves the use of data structure and algorithms to efficiently store a large amount of book information, this program allows the user to store and manage book information such as names and prices.
The problem faced is in a bookstore users require an efficient method to find a specific book. it is a time-consuming process as we have to visit the place where the book is being placed.
The project include features such as the ability to search for and sort books, as well as the ability to add, edit, and delete entries. Additionally, It may include the implementation of data structures such as linked lists, binary search,sorting and more 

DSA Features and algorithm use
Link list
Sorting algorithm
Search algorithm
File handling
Methods